Changed class signature, Updated Dictionary, and can now create and process commands
Removed sealed accessor and inheritance from Monobehaviour.

Commands Dictionary Value type from ConsoleCommand to IConsoleCommand.

Added both a generic and nongeneric version of AddCommand to create an instance of commands both with and without parameters.

Added ParseCommand method to modify input to find the correct command from the dictionary and then send a string array to the command class to process the arguments.