Added generic version of class, changed class signature, renamed fields and added the ability to invoke commands
Created a generic version of ConsoleCommand that takes in a single parameter of T.

Changed class signature to remove the sealed accessor and to implement the IConsoleCommand interface.

Rename _command field to _commandTitle

Added ProcessArgs method to handle any arguments presented. The non-generic version just calls the InvokeCommand method.
The generic version uses Convert.ChangeType to convert the string arguments to the generic type the instance is assigned.