Added support for the addcommand attribute
The command manager searches for methods with the add command attribute and gathers monobehavior instances to invoke commands.

The GetUtilityCommands() method was added to gather the utility commands like help and clear. Gets the executing assembly, the UtilityCommands type and then loops through the methods and add them as commands through the addcommand attribute.

The CreateDelegate commands creates a delegate based on the method signature using the MethodInfo and target.