Updated ConsoleCommand for attribute feature
Modified constructor parameters and fields to adhere to the new attribute feature.
Changed the command field from a Action to a Delegate and added an object field to pass in the instance. Added ParamTypes fields to grab an array of parameter types.

The InvokeCommand() method is modified to take in an object array to invoke the command.

The ProcessArgs(string[] args) method is modified to handle converting a string array into the appropriate parameter types.