Update: Supports commands for non-mono scripts
Update: Implemented RegisterType method to take in the object instance and type.
A MethodInfo array is extracted from the type and is iterated through to find methods with the AddCommandAttribute.
Once an attribute has been found, a new delegate is created reflecting the MethodInfo and a new ConsoleCommand instance is made to hold it.

Fix: Allows users to enter commands without spaces for commands with no parameters

Removed: all AddCommand methods as they are no longer being used.