Update: Centralized Invoking Command
To centralize how a command is invoked, I modified the ProcessArgs signature to make arguments an optional parameter. The args parameter receives a null check that will call the parameter-less InvokeCommand method for parameter-less methods.
Removed Constructors with a signature requiring an object instance as ConsoleObjects will keep track of the object instance.

CommandKey field added as new requirement for IConsoleCommand.