Feature: Creates CommandObjects for instance support
CommandManager has been set up to create CommandObjects to store object instances and the commands associated with them.

Commands Dictionary now stores a Value of CommandObject.
ParseCommand method now stores a instanceKey field that grabs the CommandObject and then calls the ProcessCommand method to input the commandKey and parameters.
GetAllCommands now iterates through all monobehaviours within the scene to create delegates and to assign the gameObject to store the instance.
AddCommands method generates new Commands to already existing CommandObjects or creates new ones.

ConsoleCommands are no placed in arrays to fit new parameters.

GenerateInstanceKey: generates an instance key based on either the GameObject name or the Type name. If a key already exists, then the name will be appended with a <_number>. The number is incremented in a while loop.

LastCommand is renamed to LastCMDInstance.