Feature: Command Instances
The AddCommand method will add a command to an existing CommandObject or create a new CommandObject if it does not exist. This is done by comparing gameobjects instances in memory.
AddCommand method is now called after creating a new Delegate.
Added warning logs for returning null CommandObjects.
Updated GenerateInstanceKey signature to receive a Monobehaviour. Added checks to avoid duplication.
AddCommands method was removed as it was no longer required to add collections of ConsoleCommands. CommandObjects now store commands in a list. Therefore, commands can be added individually.
Updated parameters to look up dictionary with the instanceKey instead of the commandKey.
Reduces Monobehaviour iterations: Check if an IEnumerator of Monobehaviours has any elements and then moves on to the next loop.

CommandObject: Added constructor for monobehaviours instead object instances. Changed AddCommand signature to store an instance for the command itself to store.
AddCommand now sets the monobehaviour instance for the current command and then adds the command to the commands list. GetCommand uses the List's Find method to get the command that matches the input entered.
Changed ProcessCommand method to ensure that commandKeys are entered.