Feature: Rename InstanceKeys for CommandObjects
Added a static void RenameInstanceKey method with two arguments, oldKey to find the original value and the key to replace the old one. The InstanceKey for the CommandObject is then replaced using a SetInstanceKey method that takes in a new InstanceKey.
Adds a check for GeneratingInstanceKey for objects to see if the key exists and if so returns the key as is.