Feature: ConsoleObjects are added upon spawning new objects as prefabs
The SceneTracker class checks and updates based on the current root count from the scene, grabs the new object and then passes it into the CopyCommand function.
The CopyCommand function tries to find a source object to grab from and then makes a local copy of the appropriate commands if present. It will then generate an instance key and then a new CommandObject. It then iterates through each monobehaviour instance then through each methodInfo to find the appropriate commands if necessary. Regardless all commands are added to the new CommandObject.